Next commit

+++ Jenkinsfile
@@ -1,8 +1,9 @@
 pipeline {
-    agent { label 'ubuntu'} 
+    agent none 
 
     stages {
         stage('Build') {
+            node { label 'ubuntu' }
             steps {
                 git 'https://github.com/shurasius/rabbit-proxy-project.git'
                 sh 'sudo chmod 755 create-or-update-stack.sh'
@@ -21,20 +22,22 @@ pipeline {
             steps {
                 sh './create-or-update-stack.sh us-west-2 RabbitProxyProject-v1-${BUILD_NUMBER} --template-body file://rabbitt_project_optimized.yml --parameters ParameterKey=BuildVersion,ParameterValue=rabbit_artifact_1.${BUILD_NUMBER}.tar  ParameterKey=KeyName,ParameterValue=oregon_pair_1 ParameterKey=InstanceType,ParameterValue=t2.micro ParameterKey=SSHLocation,ParameterValue=0.0.0.0/0 ParameterKey=VPC,ParameterValue=vpc-8811e8f0 ParameterKey=Subnets,ParameterValue="subnet-2dd7f377\\,subnet-560a5f2f\\,subnet-f796e1bc" --capabilities CAPABILITY_NAMED_IAM'
 
-                sh "LBDNS=`aws cloudformation --region us-west-2 describe-stacks --stack-name RabbitProxyProject-v1-83 --query 'Stacks[0].Outputs[0].OutputValue' | sed 's/\\"//g'`"
-                sh 'curl $LBDNS'
+                sh ''' export NGINXDNS=`aws cloudformation describe-stacks --stack-name rabbit-nginx-dev --query 'Stacks[0].Outputs[0].OutputValue' | cut -d'"' -f2`
+                echo $NGINXDNS 
+                if [[ `curl $NGINXDNS` && `curl $NGINXDNS/rabbit-mgmt` ]]; then exit 0; else exit 1; fi
 
                 sh 'aws cloudformation delete-stack --region us-west-2  --stack-name RabbitProxyProject-v1-${BUILD_NUMBER}'
 
             }
         }
         stage('Deploy Prod') {
+            node { label 'redhat'}
             steps {
                 sh './create-or-update-stack.sh us-west-2 RabbitProxyProject-v1-${BUILD_NUMBER} --template-body file://rabbitt_project_optimized.yml --parameters ParameterKey=BuildVersion,ParameterValue=rabbit_artifact_1.${BUILD_NUMBER}.tar  ParameterKey=KeyName,ParameterValue=oregon_pair_1 ParameterKey=InstanceType,ParameterValue=t2.micro ParameterKey=SSHLocation,ParameterValue=0.0.0.0/0 ParameterKey=VPC,ParameterValue=vpc-8811e8f0 ParameterKey=Subnets,ParameterValue="subnet-2dd7f377\\,subnet-560a5f2f\\,subnet-f796e1bc" --capabilities CAPABILITY_NAMED_IAM'
 
-                sh "LBDNS=`aws cloudformation --region us-west-2 describe-stacks --stack-name RabbitProxyProject-v1-${BUILD_NUMBER} --query 'Stacks[0].Outputs[0].OutputValue' | sed 's/\\"//g'`"  
-                sh 'echo $LBDNS'
-                sh 'curl $LBDNS'
+                sh ''' export NGINXDNS=`aws cloudformation describe-stacks --stack-name rabbit-nginx-dev --query 'Stacks[0].Outputs[0].OutputValue' | cut -d'"' -f2`
+                echo $NGINXDNS 
+                if [[ `curl $NGINXDNS` && `curl $NGINXDNS/rabbit-mgmt` ]]; then exit 0; else exit 1; fi
 
                 sh 'aws cloudformation delete-stack --region us-west-2  --stack-name RabbitProxyProject-v1-${BUILD_NUMBER}'
